Feature: Simplify response handling in EtterlatteClient by consolidating status mapping
